+++ Source/VungleAdapterRouter.swift
@@ -68,39 +68,39 @@ extension VungleAdapterRouter {
         } else if let placementID = placementID {
             recordLoadEnd(forPlacement: placementID)
         }
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: false)?.vungleAdPlayabilityUpdate?(isAdPlayable, placementID: placementID, adMarkup: adMarkup, error: partnerError)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleAdPlayabilityUpdate?(isAdPlayable, placementID: placementID, adMarkup: adMarkup, error: partnerError)
     }
     
     func vungleWillShowAd(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: false)?.vungleWillShowAd?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleWillShowAd?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleDidShowAd(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: false)?.vungleDidShowAd?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleDidShowAd?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleAdViewed(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: false)?.vungleAdViewed?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleAdViewed?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleTrackClick(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: true)?.vungleTrackClick?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleTrackClick?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleWillLeaveApplication(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: true)?.vungleWillLeaveApplication?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleWillLeaveApplication?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleWillCloseAd(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: true)?.vungleWillCloseAd?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleWillCloseAd?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleDidCloseAd(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: true)?.vungleDidCloseAd?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleDidCloseAd?(forPlacementID: placementID, adMarkup: adMarkup)
     }
     
     func vungleRewardUser(forPlacementID placementID: String?, adMarkup: String?) {
-        ad(for: placementID, adMarkup: adMarkup, prioritizeVisibleBanner: true)?.vungleRewardUser?(forPlacementID: placementID, adMarkup: adMarkup)
+        ad(for: placementID, adMarkup: adMarkup)?.vungleRewardUser?(forPlacementID: placementID, adMarkup: adMarkup)
     }
 }
 
@@ -162,7 +162,7 @@ extension VungleAdapterRouter {
 private extension VungleAdapterRouter {
     
     /// Fetches a stored ad adapter and logs an error if none is found.
-    func ad(for partnerPlacement: String?, adMarkup: String?, prioritizeVisibleBanner: Bool, functionName: StaticString = #function) -> VungleSDKHBDelegate? {
+    func ad(for partnerPlacement: String?, adMarkup: String?, functionName: StaticString = #function) -> VungleSDKHBDelegate? {
         
         func adMatchesPlacementAndMarkup(_ ad: PartnerAd) -> Bool {
             ad.request.partnerPlacement == partnerPlacement && ad.request.adm == adMarkup
@@ -175,19 +175,6 @@ private extension VungleAdapterRouter {
         guard let ad = adapter.storage.ads.last(where: adMatchesPlacementAndMarkup) as? VungleAdapterAd else {
             adapter.log("\(functionName) call ignored with placementID \(partnerPlacement), no corresponding partner ad found.")
             return nil
-        }
-        // It is possible that multiple banner ads with the same placement are created given that Chartboost Mediation preloads the next banner while the previous one is
-        // still visible as part of its auto-refresh process.
-        // Knowing which `PartnerAd` instance corresponds to each delegate method call gets tricky.
-        // Here we do a best attempt at it by forwarding calls we know should happen only on visible banners to the currently visible banner ad.
-        // Other calls are forwarded to the latest created ad.
-        // The solution is not ideal but the only one to make this work without preventing multiple loads for ads with the same placement, which would make
-        // Chartboost Mediation banner auto-refresh stop if the same partner placement won over and over again.
-        // It should work fine under the assumption that Chartboost Mediation does not support the creation of multiple banners with the same placement by publishers, which means
-        // at a single point in time there should be at most two Vungle banners with the same placement loaded: the visible one and the preloaded one intended to replace it.
-        if ad.request.format == .banner && prioritizeVisibleBanner {
-            let firstVisibleBanner = adapter.storage.ads.first(where: { adMatchesPlacementAndMarkup($0) && $0.inlineView?.window != nil }) as? VungleAdapterAd
-            return firstVisibleBanner ?? ad
         }
         return ad
     }