+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 Note the first digit of every adapter version corresponds to the major version of the Chartboost Mediation SDK compatible with that adapter. 
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
+### 4.6.12.1.1
+- Removed support for loading multiple banners for the same placement.
+- This version of the adapters has been certified with VungleSDK-iOS 6.12.1.
+
 ### 4.6.12.1.0
 - This version of the adapters has been certified with VungleSDK-iOS 6.12.1.